externalize msd_pyramid add export wrappers
