+++ modules/xfeatures2d/include/opencv2/xfeatures2d.hpp
@@ -901,7 +901,7 @@ public:
  * The interface is equivalent to @ref Feature2D, adding operations for
  * @ref Elliptic_KeyPoint "Elliptic_KeyPoints" instead of @ref KeyPoint "KeyPoints".
  */
-class CV_EXPORTS AffineFeature2D : public Feature2D
+class CV_EXPORTS_W AffineFeature2D : public Feature2D
 {
 public:
     /**
@@ -962,7 +962,7 @@ Features are ellipses (similar to MSER, however a MSER feature can never be a
 TBMR feature and vice versa).
 
 */
-class CV_EXPORTS TBMR : public AffineFeature2D
+class CV_EXPORTS_W TBMR : public AffineFeature2D
 {
 public:
     CV_WRAP static Ptr<TBMR> create(int min_area = 60,
