+++ modules/xfeatures2d/src/tbmr.cpp
@@ -3,77 +3,12 @@
 // directory of this distribution and at http://opencv.org/license.html.
 
 #include "precomp.hpp"
+#include "msd_pyramid.hpp"
 
 namespace cv
 {
 namespace xfeatures2d
 {
-/*!
-MSD Image Pyramid. (from msd.cpp)
-*/
-class TBMRImagePyramid
-{
-    // Multi-threaded construction of the scale-space pyramid
-    struct TBMRImagePyramidBuilder : ParallelLoopBody
-    {
-        TBMRImagePyramidBuilder(const cv::Mat &_im,
-                                std::vector<cv::Mat> *_m_imPyr,
-                                float _scaleFactor)
-        {
-            im = &_im;
-            m_imPyr = _m_imPyr;
-            scaleFactor = _scaleFactor;
-        }
-
-        void operator()(const Range &range) const CV_OVERRIDE
-        {
-            for (int lvl = range.start; lvl < range.end; lvl++)
-            {
-                float scale = 1 / std::pow(scaleFactor, (float)lvl);
-                (*m_imPyr)[lvl] = cv::Mat(cv::Size(cvRound(im->cols * scale),
-                                                   cvRound(im->rows * scale)),
-                                          im->type());
-                cv::resize(*im, (*m_imPyr)[lvl],
-                           cv::Size((*m_imPyr)[lvl].cols, (*m_imPyr)[lvl].rows),
-                           0.0, 0.0, cv::INTER_AREA);
-            }
-        }
-        const cv::Mat *im;
-        std::vector<cv::Mat> *m_imPyr;
-        float scaleFactor;
-    };
-
-  public:
-    TBMRImagePyramid(const cv::Mat &im, const int nLevels,
-                     const float scaleFactor = 1.6f);
-    ~TBMRImagePyramid();
-
-    const std::vector<cv::Mat> getImPyr() const { return m_imPyr; };
-
-  private:
-    std::vector<cv::Mat> m_imPyr;
-    int m_nLevels;
-    float m_scaleFactor;
-};
-
-TBMRImagePyramid::TBMRImagePyramid(const cv::Mat &im, const int nLevels,
-                                   const float scaleFactor)
-{
-    m_nLevels = nLevels;
-    m_scaleFactor = scaleFactor;
-    m_imPyr.clear();
-    m_imPyr.resize(nLevels);
-
-    m_imPyr[0] = im.clone();
-
-    if (m_nLevels > 1)
-    {
-        parallel_for_(Range(1, nLevels),
-                      TBMRImagePyramidBuilder(im, &m_imPyr, scaleFactor));
-    }
-}
-
-TBMRImagePyramid::~TBMRImagePyramid() {}
 
 class TBMR_Impl CV_FINAL : public TBMR
 {
@@ -576,7 +511,7 @@ void TBMR_Impl::detect(InputArray _image,
     float *dupl_ptr = dupl.ptr<float>();
 
     std::vector<Mat> pyr;
-    TBMRImagePyramid scaleSpacer(src, m_cur_n_scales, m_scale_factor);
+    MSDImagePyramid scaleSpacer(src, m_cur_n_scales, m_scale_factor);
     pyr = scaleSpacer.getImPyr();
 
     int oct = 0;
